move directories and check url structure
