+++ fibre/grpc/set_getter_test.go
@@ -1,10 +1,10 @@
-package validator_test
+package grpc_test
 
 import (
 	"context"
 	"testing"
 
-	"github.com/celestiaorg/celestia-app/v6/fibre/validator"
+	fibregrpc "github.com/celestiaorg/celestia-app/v6/fibre/grpc"
 	cmted25519 "github.com/cometbft/cometbft/crypto/ed25519"
 	coregrpc "github.com/cometbft/cometbft/rpc/grpc"
 	core "github.com/cometbft/cometbft/types"
@@ -32,7 +32,7 @@ func TestGrpcGetter_GetByHeight_Success(t *testing.T) {
 	}
 
 	// Create getter and fetch validator set
-	getter := validator.NewGrpcGetter(mockClient)
+	getter := fibregrpc.NewGrpcGetter(mockClient)
 	result, err := getter.GetByHeight(context.Background(), testHeight)
 
 	// Verify results
@@ -46,7 +46,7 @@ func TestGrpcGetter_GetByHeight_Success(t *testing.T) {
 func TestGrpcGetter_GetByHeight_ZeroHeight(t *testing.T) {
 	// Test that GetByHeight rejects zero height
 	var client coregrpc.BlockAPIClient
-	getter := validator.NewGrpcGetter(client)
+	getter := fibregrpc.NewGrpcGetter(client)
 
 	_, err := getter.GetByHeight(context.Background(), 0)
 	if err == nil {
